Responsive design tweaks
Reworked the container hier on the mobile first foundation 6 skin.
Feel like I am getting the hang of their grid finally.
There are a few parts that are nastily coded inside the lisp for now,
where it relies on injecting HTML with specific class dependencies,
that hopefully can be tidied up with a later refactor