changes to fix header -> content transition
hopefully deal with blank lines a little better.
There is a bug with empty lines in content. I think it is because the
logic for measuring the transition from headers to body was faulty,
and would flip any time there was a blank line. this patch changes
read post to account for the state of the headers when checking to see
if there's a transition (i.e. don't consider header processing as
potentially finished until there is at least one header), and inverts
the state tracking to use a 'headers finished' flag, not a 'i'm doing
headers flag' , which may be easier to read.

I think this needs to be reworked to be a more obvious state machine