adds cache support
wrap index serving in a filesystem cache built using clache
sets up some new configuration vars ( cache dir , and cache instance )
we now set a cache-version key at boot to the mtime of the source
document directory.

there's a new cache-version util function that retrieves this value -
internally comparing it to the mtime on the directory. If the
directory entries have changed, it purges the cache and updates the
version key accordingly.

build index now uses a clache wrapper that calls cache-version whilst
computing a key.